+++ tests/testthat/tests.brmsfit-methods.R
@@ -72,7 +72,6 @@ test_that("all S3 methods have reasonable ouputs", {
   expect_equal(dim(R2), c(nsamples(fit1), 1))
   R2 <- bayes_R2(fit2, newdata = model.frame(fit2)[1:5, ], re_formula = NA)
   expect_equal(dim(R2), c(1, 4))
-  expect_error(bayes_R2(fit4), "'bayes_R2' is not defined for ordinal")
   R2 <- bayes_R2(fit6)
   expect_equal(dim(R2), c(2, 4))